+++ App/Course App/Course App/SharedUI/DesignSystem/FontSize.swift
@@ -1,0 +1,14 @@
+//
+//  FontSize.swift
+//  Course App
+//
+//  Created by Patrik Urban on 03.06.2024.
+//
+
+import Foundation
+enum FontSize: CGFloat {
+    case size36 = 36
+    case size28 = 28
+    case size20 = 20
+    case size12 = 12
+}
